+++ realsense_camera/include/realsense_camera/r200_nodelet.h
@@ -34,8 +34,8 @@
 
 #include <dynamic_reconfigure/server.h>
 
-#include "realsense_camera/r200_paramsConfig.h"
-#include "base_nodelet.h"
+#include <realsense_camera/r200_paramsConfig.h>
+#include <realsense_camera/base_nodelet.h>
 
 namespace realsense_camera
 {
