+++ realsense_camera/src/r200_nodelet.cpp
@@ -28,7 +28,7 @@
  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  *******************************************************************************/
 
-#include "r200_nodelet.h"
+#include <realsense_camera/r200_nodelet.h>
 
 PLUGINLIB_EXPORT_CLASS (realsense_camera::R200Nodelet, nodelet::Nodelet)
