+++ realsense_camera/test/camera_core.cpp
@@ -27,7 +27,7 @@
  OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
  OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  *******************************************************************************/
-#include "gtest/gtest.h"
+#include <gtest/gtest.h>
 #include "camera_core.h"
 
 using namespace std;
