+++ realsense_camera/test/camera_core.h
@@ -40,19 +40,19 @@
 #include <image_transport/image_transport.h>
 
 #include <camera_info_manager/camera_info_manager.h>
-#include "std_msgs/String.h"
+#include <std_msgs/String.h>
 #include <ros/ros.h>
 #include <pcl/conversions.h>
 #include <pcl/point_cloud.h>
 #include <pcl/point_types.h>
 #include <pcl/PCLPointCloud2.h>
 #include <pcl_conversions/pcl_conversions.h>
-#include "std_msgs/Float32MultiArray.h"
+#include <std_msgs/Float32MultiArray.h>
 #include <cv_bridge/cv_bridge.h>
 #include <realsense_camera/cameraConfiguration.h>
 #include <tf/transform_listener.h>
 #include <librealsense/rs.h>
-#include "constants.h"
+#include <realsense_camera/constants.h>
 
 using namespace realsense_camera;
